[FEATURE] Ajax ESI through Request/Dispatch Widget
Ajax mode can now be enabled in the Request/Dispatch Widget. Depends on
inclusion of new static TypoScript to enable the special typeNum and
(not included) jQuery library to perform the Ajax request itself.

Note that currently, plugins which require FlexForm settings in order
to function (and do not have fallbacks or TypoScript options to
configure those parameters) are not currently supported. A future
revision will solve this by allowing the contentObject UID to be
configured and that contentObject's record data (including FlexForm) to
be used when handling the Request through the Ajax endpoint.